refactor(cli): rename peer-polling.ts → peer-wait.ts
The file no longer polls — commit 1c3ded158 replaced the 100ms loop with
awareness.observe() subscriptions. The name was actively misleading.
